requirements.txt
databricks-sql-connector==3.4.0
pandas==2.2.3
python-dotenv==1.0.1
pyspark==3.5.3
boto3==1.34.95
botocore==1.34.95